requirements.txt
# Core dependencies for VisionPivot Attendance System
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
facenet-pytorch==2.5.3
firebase-admin==6.2.0
pyttsx3==2.90
PyQt5==5.15.9
yagmail==0.15.293
pandas==2.1.1
numpy==1.24.3
Pillow==10.0.1

# Additional dependencies that might be needed
scipy==1.11.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
openpyxl==3.1.2
